Simple script to check the beanstalkd `ready` jobs in all tubes
To run: `ruby jobs.rb beanstalk://127.0.0.1/ 25`

This will check the `ready` jobs on all tubes. If any of them are over the
`MAX_JOBS` threshold, the script will return 1.

If no tubes have more than `MAX_JOBS` jobs, the script will return 0.

If there is an error with the script, it will return 3.

By default, `MAX_JOBS` is 25 and the beanstalkd url is `beanstalk://127.0.0.1/`